Adding the user, authentication and authorization

+++ app/controllers/comments_controller.rb
@@ -1,9 +1,11 @@
 class CommentsController < ApplicationController
     before_action :find_post
-
+    before_action :authenticated_user!
+    
     def create
         @comment = Comment.new(params.require(:comment).permit(:body))
         @comment.post = @post
+        @comment.user = current_user
 
         if @comment.save
             flash[:notice] = "Comment created successfully!"
@@ -18,17 +20,19 @@ class CommentsController < ApplicationController
 
     def destroy
         @comment = Comment.find(params[:id])
-
-        if @comment.destroy
-            flash[:notice] = "Comment deleted successfully!"
+        if can?(:destroy, @comment)
+            if @comment.destroy
+                flash[:notice] = "Comment deleted successfully!"
+            else
+                flash[:alert] = "Cannot deleted comment!"
+            end
+            redirect_to post_path(@post)
         else
-            flash[:alert] = "Cannot deleted comment!"
+            redirect_to post_path(@post), alert: "Not authorized"
         end
-        redirect_to post_path(@post)
     end
 
     private
-
     def find_post
         @post = Post.find(params[:post_id])
     end
